Restructure Akka runtime deployment code
Extract some commonalities into a trait and clarify the flow
