+++ runtime/src/main/scala/endless/runtime/akka/deploy/Deployer.scala
@@ -1,25 +1,24 @@
-package endless.runtime.akka
+package endless.runtime.akka.deploy
 
-import akka.actor.typed.scaladsl.Behaviors
 import akka.actor.typed.{ActorRef, ActorSystem, Behavior}
 import akka.cluster.sharding.typed.ShardingEnvelope
-import akka.cluster.sharding.typed.scaladsl.{ClusterSharding, EntityContext, EntityTypeKey}
-import akka.persistence.typed.scaladsl.{Effect, EventSourcedBehavior}
-import akka.persistence.typed.{PersistenceId, RecoveryCompleted, RecoveryFailed}
+import akka.cluster.sharding.typed.scaladsl.{ClusterSharding, EntityContext}
+import akka.persistence.typed.scaladsl.EventSourcedBehavior
 import akka.util.Timeout
 import cats.effect.kernel.{Async, Resource}
-import cats.effect.std.Dispatcher
 import cats.syntax.applicative._
 import cats.syntax.flatMap._
 import cats.syntax.functor._
-import cats.syntax.show._
 import cats.tagless.FunctorK
 import endless.core.entity._
 import endless.core.event.EventApplier
 import endless.core.interpret.EffectorT._
 import endless.core.interpret._
 import endless.core.protocol.{CommandProtocol, CommandRouter, EntityIDCodec}
+import endless.runtime.akka.ShardingCommandRouter
 import endless.runtime.akka.data._
+import endless.runtime.akka.deploy.Deployer.EffectorParameters
+import endless.runtime.akka.deploy.internal.EventSourcedShardedEntityDeployer
 import org.typelevel.log4cats.Logger
 
 trait Deployer {
@@ -134,9 +133,8 @@ trait Deployer {
       customizeEntity
     )
 
-  /** Overload of [[Deployer.deployEntity]] that accepts creation functions expressed in `F`
-    * context. This is particularly useful for creating a stateful effector (which is specific to an
-    * entity).
+  /** Overload of [[deployEntity]] that accepts creation functions expressed in `F` context. This is
+    * particularly useful for creating a stateful effector (which is specific to an entity).
     */
   def deployEntityF[F[_]: Async: Logger, S, E, ID: EntityIDCodec, Alg[
       _[_]
@@ -166,140 +164,21 @@ trait Deployer {
   ): Resource[F, (RepositoryAlg[F], ActorRef[ShardingEnvelope[Command]])] = {
     implicit val commandRouter: CommandRouter[F, ID] = ShardingCommandRouter.apply
     val repositoryT = RepositoryT.apply[F, ID, Alg]
-    Resource.eval(
-      for {
-        interpretedEntityAlg <- createEntity(EntityT.instance)
-        repository <- createRepository(repositoryT)
-      } yield new DeployEntity(
+    for {
+      interpretedEntityAlg <- Resource.eval(createEntity(EntityT.instance))
+      repository <- Resource.eval(createRepository(repositoryT))
+      entity <- new EventSourcedShardedEntityDeployer(
         interpretedEntityAlg,
         repository,
         repositoryT,
         createEffector,
-        customizeBehavior,
-        customizeEntity
-      ).apply
-    )
-  }.flatten
-
-  final class EventApplierException(error: String) extends RuntimeException(error)
-
-  private type EffectorParameters[F[_], S, Alg[_[_]], RepositoryAlg[_[_]]] =
-    (Effector[EffectorT[F, S, Alg, *], S, Alg], RepositoryAlg[F], Alg[F])
-
-  private class DeployEntity[F[_]: Async: Logger, S, E, ID: EntityIDCodec, Alg[
-      _[_]
-  ]: FunctorK, RepositoryAlg[_[_]]](
-      interpretedEntityAlg: Alg[EntityT[F, S, E, *]],
-      repository: RepositoryAlg[F],
-      repositoryT: RepositoryT[F, ID, Alg],
-      createEffector: EffectorParameters[F, S, Alg, RepositoryAlg] => F[EffectorT[F, S, Alg, Unit]],
-      customizeBehavior: (
-          EntityContext[Command],
-          EventSourcedBehavior[Command, E, Option[S]]
-      ) => Behavior[Command],
-      customizeEntity: akka.cluster.sharding.typed.scaladsl.Entity[Command, ShardingEnvelope[
-        Command
-      ]] => akka.cluster.sharding.typed.scaladsl.Entity[Command, ShardingEnvelope[Command]]
-  )(implicit
-      sharding: ClusterSharding,
-      actorSystem: ActorSystem[_],
-      nameProvider: EntityNameProvider[ID],
-      commandProtocol: CommandProtocol[Alg],
-      eventApplier: EventApplier[S, E],
-      askTimeout: Timeout
-  ) {
-    private val entityTypeKey = EntityTypeKey[Command](nameProvider())
-
-    def apply: Resource[F, (RepositoryAlg[F], ActorRef[ShardingEnvelope[Command]])] =
-      Dispatcher.parallel[F].map { implicit dispatcher =>
-        val akkaEntity = akka.cluster.sharding.typed.scaladsl.Entity(
-          EntityTypeKey[Command](nameProvider())
-        ) { implicit context =>
-          Behaviors.setup { implicit actor =>
-            implicit val passivator: EntityPassivator[F] =
-              dispatcher.unsafeRunSync(EntityPassivator[F])
-            implicit val entity: Alg[F] =
-              repositoryT.entityFor(implicitly[EntityIDCodec[ID]].decode(context.entityId))
-            implicit val effector: EffectorT[F, S, Alg, Unit] =
-              dispatcher.unsafeRunSync(createEffector(EffectorT.instance, repository, entity))
-            customizeBehavior(
-              context,
-              EventSourcedBehavior
-                .withEnforcedReplies[Command, E, Option[S]](
-                  PersistenceId(entityTypeKey.name, context.entityId),
-                  Option.empty[S],
-                  commandHandler = handleCommand,
-                  eventHandler = handleEvent
-                )
-                .receiveSignal {
-                  case (state, RecoveryCompleted) =>
-                    dispatcher.unsafeRunAndForget(
-                      Logger[F].info(
-                        show"Recovery of ${nameProvider()} entity ${context.entityId} completed"
-                      ) >> effector.runS(state, entity).flatMap(passivator.apply)
-                    )
-                  case (_, RecoveryFailed(failure)) =>
-                    dispatcher.unsafeRunSync(
-                      Logger[F].warn(
-                        show"Recovery of ${nameProvider()} entity ${context.entityId} failed with error ${failure.getMessage}"
-                      )
-                    )
-                }
-            )
-          }
-        }
-        (repository, sharding.init(customizeEntity(akkaEntity)))
-      }
-
-    private def handleEvent(state: Option[S], event: E)(implicit
-        entity: Alg[F],
-        dispatcher: Dispatcher[F],
-        passivator: EntityPassivator[F],
-        effector: EffectorT[F, S, Alg, Unit]
-    ) = eventApplier.apply(state, event) match {
-      case Left(error) =>
-        dispatcher.unsafeRunSync(Logger[F].warn(error))
-        throw new EventApplierException(error)
-      case Right(newState) => newState
-    }
-
-    private def handleCommand(state: Option[S], command: Command)(implicit
-        entity: Alg[F],
-        dispatcher: Dispatcher[F],
-        passivator: EntityPassivator[F],
-        effector: EffectorT[F, S, Alg, Unit]
-    ) = {
-      val incomingCommand =
-        commandProtocol.server[EntityT[F, S, E, *]].decode(command.payload)
-      val effect = Logger[F].debug(
-        show"Handling command for ${nameProvider()} entity ${command.id}"
-      ) >> incomingCommand
-        .runWith(interpretedEntityAlg)
-        .run(state)
-        .flatMap {
-          case Left(error) =>
-            Logger[F].warn(error) >> Effect.unhandled[E, Option[S]].thenNoReply().pure[F]
-          case Right((events, reply)) if events.nonEmpty =>
-            Effect
-              .persist(events.toList)
-              .thenRun((state: Option[S]) =>
-                // run the effector asynchronously, as it can describe long-running processes
-                dispatcher
-                  .unsafeRunAndForget(effector.runS(state, entity).flatMap(passivator.apply))
-              )
-              .thenReply(command.replyTo) { _: Option[S] =>
-                Reply(incomingCommand.replyEncoder.encode(reply))
-              }
-              .pure[F]
-          case Right((_, reply)) =>
-            Effect
-              .reply[Reply, E, Option[S]](command.replyTo)(
-                Reply(incomingCommand.replyEncoder.encode(reply))
-              )
-              .pure[F]
-        }
-      dispatcher.unsafeRunSync(effect)
-    }
+        customizeBehavior
+      ).deployShardedEntity(sharding, customizeEntity).map((repository, _))
+    } yield entity
   }
+}
 
+object Deployer {
+  type EffectorParameters[F[_], S, Alg[_[_]], RepositoryAlg[_[_]]] =
+    (Effector[EffectorT[F, S, Alg, *], S, Alg], RepositoryAlg[F], Alg[F])
 }
