+++ runtime/src/main/scala/endless/runtime/akka/syntax/package.scala
@@ -1,6 +1,7 @@
 package endless.runtime.akka
 
 import endless.core.interpret.LoggerLiftingHelper
+import endless.runtime.akka.deploy.{Deployer, DurableDeployer}
 
 package object syntax {
   object deploy extends Deployer with DurableDeployer with LoggerLiftingHelper
